RawImageManipConfig: move Point2f,Size2f,RotatedRect to common
